requirements.txt
python-dotenv
ccxt
scikit-learn
numpy
google-cloud-storage
pandas
TA-lib-bin
lightgbm